send enriched clicks to clickhouse

+++ links-analytics-application/src/main/kotlin/dev/ykvlv/links/analytics/application/data/dto/IpRange.kt
@@ -1,9 +1,9 @@
 package dev.ykvlv.links.analytics.application.data.dto
 
 data class IpRange(
-    val start: Long,
-    val end: Long,
-    val asn: Long,
+    val start: UInt,
+    val end: UInt,
+    val asn: UInt,
     val country: String,
     val asnOrg: String
 )
